requirements.txt
Django==5.1.4
djangorestframework==3.15.2
Pillow==10.4.0
requests==2.32.3
python-dotenv==1.0.1
diffusers==0.31.0
transformers==4.47.1
torch==2.5.1
accelerate==1.2.1
gunicorn==23.0.0
whitenoise==6.8.2
